Refactor(#99): 코드 리팩토링 및 전역상태로 성능 최적화
